+++ app/Http/Controllers/Api/AreaSubLocalityController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\Api;
 use App\Models\area_sub_locality;
 use App\Models\area_locality;
 use Illuminate\Http\Request;
+use App\Models\area_transaction;
+use Auth;
 use App\Http\Controllers\Controller;
 
 class AreaSubLocalityController extends Controller
@@ -38,9 +40,21 @@ class AreaSubLocalityController extends Controller
             'status' => $request->status,
             ];
         area_sub_locality:: create($locality_data);
-        return response() -> json([
-                'message' => 'Sub-Locality Created',
-            ]);
+         $create_data=area_sub_locality::select('sub_locality_id','sub_locality','locality_id','status')->where('status', '1')->orderBy('sub_locality_id', 'desc')->first();
+         if($create_data){
+             $transaction_data=[
+                'table_name'=>'area_sub_localities',
+                'old_column'=> NULL,
+                'new_column'=>json_encode($create_data),
+                'action'=>'Create',
+                'updated_user'=> Auth::user()->id
+             ];
+             area_transaction::create($transaction_data);
+            
+            return response() -> json([
+                    'message' => 'Sub-Locality Created',
+                ]);
+        }
 
         }catch(\Exception $e) {
             return $this->getExceptionResponse1($e);
@@ -53,13 +67,22 @@ class AreaSubLocalityController extends Controller
             $request -> validate([
                     'sub_locality_id' => 'required|integer'
                 ]);
-            $data= area_sub_locality::select('status')->where('sub_locality_id', $request->sub_locality_id)->first();
+            $data= area_sub_locality::select('sub_locality_id','sub_locality','locality_id','status')->where('sub_locality_id', $request->sub_locality_id)->first();
             // return $data;
             if($data['status']=='1'){
                 area_sub_locality::where('sub_locality_id', $request->sub_locality_id)->update(['status' =>'0']);
             }else{
              area_sub_locality::where('sub_locality_id',$request->sub_locality_id)->update(['status' =>'1']);
             }
+             $new_data= area_sub_locality::select('sub_locality_id','sub_locality','locality_id','status')->where('sub_locality_id', $request->sub_locality_id)->first();
+              $transaction_data=[
+                    'table_name'=>'area_sub_localities',
+                    'old_column'=> json_encode($data),
+                    'new_column'=>json_encode($new_data),
+                    'action'=>'Status Update',
+                    'updated_user'=> Auth::user()->id
+                 ];
+                 area_transaction::create($transaction_data);
             return response()->json([
                 'message' => 'Sub-Locality Status Changes',
                 'status'=> 200
@@ -203,11 +226,24 @@ class AreaSubLocalityController extends Controller
         ]);
 
       try{
+        $old_data=area_sub_locality::select('sub_locality_id','sub_locality','locality_id','status')->where('sub_locality_id', $request->sub_locality_id)->first();
        area_sub_locality::where('sub_locality_id', $request->sub_locality_id)->update(['locality_id' => $request->locality_id,'status'=> $request->status,'sub_locality'=> $request->sub_locality]);
+         $new_data=area_sub_locality::select('sub_locality_id','sub_locality','locality_id','status')->where('sub_locality_id', $request->sub_locality_id)->first();
+
+             if($old_data){
+                 $transaction_data=[
+                    'table_name'=>'area_sub_localities',
+                    'old_column'=> json_encode($old_data),
+                    'new_column'=>json_encode($new_data),
+                    'action'=>'Update',
+                    'updated_user'=> Auth::user()->id
+                 ];
+                 area_transaction::create($transaction_data);
               return response()->json([
               'message' =>'data updated',
               'status'=>201
-            ], 201);
+            ], 201); 
+           }
         }catch(\Exception $e) {
               return $this->getExceptionResponse($e);
         } 
@@ -216,10 +252,23 @@ class AreaSubLocalityController extends Controller
     public function delete(Request $request)
     {
         try{
-            area_sub_locality::where('sub_locality_id', $request['sub_locality_id'])->delete();  
-            return response()->json([
-                'message' => 'deleted Successfully',
-            ], 201);
+             $data=area_sub_locality::select('sub_locality_id','sub_locality','locality_id','status')->where('sub_locality_id', $request['sub_locality_id'])->first();
+             if($data){
+                 $transaction_data=[
+                    'table_name'=>'area_sub_localities',
+                    'old_column'=> json_encode($data),
+                    'new_column'=>NULL,
+                    'action'=>'Delete',
+                    'updated_user'=> Auth::user()->id
+                 ];
+                 area_transaction::create($transaction_data);
+                 area_sub_locality::where('sub_locality_id', $request['sub_locality_id'])->delete();  
+                return response()->json([
+                    'message' => 'deleted Successfully',
+                ], 201);
+
+             }
+           
         }catch(\Exception $e) {
             return $this->getExceptionResponse($e);
         }
